feat: 混动UI完整实现 + display_manager重构 + QML组件新增
混动专用配置（HYBRID-001~005）:
- config/can_ids.yaml: 新增16个混动CAN信号（充电/电池/电机）
- config/alarm_rules.yaml: 新增电系报警（过压/欠压/过流/高温）
- config/indicators.yaml: 新增能量流图指示灯
- config/can_signal_status.yaml: BMS信号健康监控

UI组件新增:
- src/ui/AlarmBanner.qml: 报警横幅（多语言支持）
- src/ui/AlarmBannerItem.qml: 单条报警项
- src/ui/EnergyFlowDiagram.qml: 能量流示意图
- src/ui/MotorRPMGauge.qml: 电机转速表
- src/ui/SpeedGauge.qml: 车速表

C++ Layer2重构:
- src/layer2/display_manager.cpp/h: 新增DisplayManager统一管理显示数据
- src/layer2/can_signal_monitor.cpp/h: 信号健康监控重构
- can-processor/main.cpp: 混动数据处理逻辑

QML绑定修复:
- DashboardMain.qml: 混动模式能量流 + AlarmBanner集成
- GaugeCanvas.qml: Canvas表盘绘制完善
- IndicatorLight.qml: 多状态指示灯
- SeatBeltZone.qml: 安全带状态显示

需求文档更新:
- REQ-ALM-005: 胎压低报警（更新测试用例）
- REQ-IND-006~012: 指示灯需求完善

技术约束:
- processFrame() DisplayData写入（switch-case by index）
- byte_start >= len 边界检查
- AlarmRuntime/IndicatorRuntime 正确注册EventBus

+++ src/layer2/can_signal_monitor.h
@@ -22,7 +22,7 @@ struct SignalMonitorDef {
     float       max_value;
     float       max_delta;      // 0=不检测突变
     bool        smoothing;
-    int         smoothing_window;
+    uint8_t     smoothing_window;
 };
 
 struct SignalState {
